Fix long zero point issue.

+++ mqbench/observer.py
@@ -523,7 +523,7 @@ class MSEObserver(ObserverBase):
             new_max = x_max * (1.0 - (i * 0.01))
             scale, zero_point = self._calculate_qparams(new_min, new_max)
             x_q = torch.fake_quantize_per_channel_affine(
-                x, scale, zero_point.long() if _version_under_1100 else zero_point, ch_axis, 
+                x, scale, zero_point.int() if _version_under_1100 else zero_point.int(), ch_axis, 
                 self.quant_min, self.quant_max)
             score = self.lp_loss(x_q, x, reduce_dim)
             update_idx = (score < best_score)
@@ -602,7 +602,7 @@ class EMAMSEObserver(ObserverBase):
             new_max = x_max * (1.0 - (i * 0.01))
             scale, zero_point = self._calculate_qparams(new_min, new_max)
             x_q = torch.fake_quantize_per_channel_affine(
-                x, scale, zero_point.long() if _version_under_1100 else zero_point, ch_axis, 
+                x, scale, zero_point.int() if _version_under_1100 else zero_point.int(), ch_axis, 
                 self.quant_min, self.quant_max)
             score = self.lp_loss(x_q, x, reduce_dim)
             update_idx = (score < best_score)
